refactor: Remove instanceof operator
